Move SecurityScheme annotation to model mapping to core
This moves the SecurityParser class responsible for populating
SecurityScheme models from SecurityScheme annotations to the core
module.
The move allows using it without taking the dependency on JAX-RS.
